[Refactor] 팀 평가 대시보드 수정
